moved GpumemTensor to driver.hpp
