+++ driver/bn_driver.hpp
@@ -63,143 +63,6 @@
 
 #define MIO_DRIVER_BN_REFERENCE_COMPUTE_3D_AS_2D 1 // Resolves issue #1974
 
-//========================
-
-template <typename Tgpu>
-class GpumemTensor
-{
-    std::unique_ptr<GPUMem> dev;
-    tensor<Tgpu> host;
-    bool is_gpualloc = false;
-
-public:
-    void SetGpuallocMode(bool v) { is_gpualloc = v; }
-    tensor<Tgpu>& GetTensor() { return host; }
-
-    void AllocOnHost(miopenTensorDescriptor_t t)
-    {
-        host = tensor<Tgpu>(miopen::deref(t));
-        if(is_gpualloc) // We do not need host data.
-        {
-            host.data.clear();
-            host.data.shrink_to_fit(); // To free host memory.
-        }
-    }
-    template <typename T>
-    void AllocOnHost(tensor<T> t)
-    {
-        AllocOnHost(&t.desc);
-    }
-
-    std::vector<Tgpu>& GetVector()
-    {
-        if(is_gpualloc)
-            MIOPEN_THROW("[MIOpenDriver] GpumemTensor::GetVector should not be called in "
-                         "'--gpualloc 1' mode");
-        return host.data;
-    }
-
-    Tgpu* GetVectorData() { return is_gpualloc ? nullptr : host.data.data(); }
-    std::size_t GetVectorSize() const { return is_gpualloc ? 0 : host.data.size(); }
-
-    void
-    InitHostData(const size_t sz,     //
-                 const bool do_write, // If set to false, then only generate random data. This is
-                                      // necessary to reproduce values in input buffers even if some
-                                      // directions are skipped. For example, inputs for Backward
-                                      // will be the same for both "-F 0" and "-F 2".
-                 std::function<Tgpu()> generator)
-    {
-        if(is_gpualloc)
-        {
-            /// In gpualloc mode, we do not care about reproducibility of results, because
-            /// validation is not used. Therefore, we do not have to always generate random value
-            /// (\ref move_rand)
-            return;
-        }
-
-        for(size_t i = 0; i < sz; ++i)
-        {
-            /// \anchor move_rand
-            /// Generate random value, even if buffer is unused. This provides the same
-            /// initialization of input buffers regardless of which kinds of
-            /// convolutions are currently selectedfor testing (see the "-F" option).
-            /// Verification cache would be broken otherwise.
-            auto val = generator();
-            if(do_write)
-                GetVector()[i] = val;
-        }
-    }
-
-    status_t AllocOnDevice(stream, context_t ctx, const size_t sz)
-    {
-        dev = std::make_unique<GPUMem>(ctx, sz, sizeof(Tgpu));
-        return STATUS_SUCCESS;
-    }
-
-    status_t AllocOnDeviceAndInit(stream q, context_t ctx, const size_t sz)
-    {
-        AllocOnDevice(q, ctx, sz);
-        if(is_gpualloc)
-        {
-            /// \anchor gpualloc_random_init
-            /// In gpualloc mode, we do not want to leave input buffers uninitialized, because
-            /// there could be NaNs and Infs, which may affect the performance (which we are
-            /// interested to evaluate in this mode). Initialization with all 0's is not the
-            /// best choice as well, because GPU HW may optimize out computations with 0's and
-            /// that could affect performance of kernels too. That is why we are using
-            /// rocrand to initialize input buffers.
-            ///
-            /// However we do not care about precision in gpualloc mode, because validation
-            /// is not used. Therefore, range (0,1] is fine.
-            return gpumemrand::gen_0_1(static_cast<Tgpu*>(GetDevicePtr()), sz);
-        }
-        return dev->ToGPU(q, GetVectorData());
-    }
-
-    template <typename T>
-    status_t AllocOnDevice(stream, context_t ctx, const size_t sz, std::vector<T>&)
-    {
-        static_assert(std::is_same<T, float>::value           //
-                          || std::is_same<T, int32_t>::value, //
-                      "Before enabling more types, check thoroughly.");
-        dev = std::make_unique<GPUMem>(ctx, sz, sizeof(T));
-        return STATUS_SUCCESS;
-    }
-
-    template <typename T>
-    status_t AllocOnDeviceAndInit(stream q, context_t ctx, const size_t sz, std::vector<T>& init)
-    {
-        AllocOnDevice(q, ctx, sz, init);
-        if(is_gpualloc)
-        {
-            /// \ref gpualloc_random_init
-            return gpumemrand::gen_0_1(static_cast<Tgpu*>(GetDevicePtr()), sz);
-        }
-        return dev->ToGPU(q, init.data());
-    }
-
-    status_t CopyFromDeviceToHost(stream q)
-    {
-        return is_gpualloc ? STATUS_SUCCESS : dev->FromGPU(q, GetVectorData());
-    }
-
-    template <typename T>
-    status_t CopyFromDeviceToHost(stream q, tensor<T>& t)
-    {
-        return is_gpualloc ? STATUS_SUCCESS : dev->FromGPU(q, t.data.data());
-    }
-
-    template <typename T>
-    status_t CopyFromDeviceToHost(stream q, std::vector<T>& v)
-    {
-        return is_gpualloc ? STATUS_SUCCESS : dev->FromGPU(q, v.data());
-    }
-
-    auto GetDevicePtr() -> auto { return dev->GetMem(); }
-};
-//========================
-
 //#define BN_RUNFOR_PROFILER
 
 template <typename Tgpu, typename Tref, typename Tmix = Tgpu>
